#450 fix all features in bplan
